Add confirmation message while trying to delete a user.

+++ resources/views/users/actions.blade.php
@@ -2,10 +2,9 @@
 	<div class="row">
 		<div class="col-md-12">
 			<div>
-				{{ Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $user->id]]) }}
+				{{ Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $user->id], 'onsubmit' => 'return confirm("are you sure you want to delete this user ?")']) }}
 				    {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
 				{{ Form::close() }}
-				{{-- <a href="#" class="btn btn-danger"><span class="glyphicon glyphicon-trash" aria-hidden="true"></span> Delete</a> --}}
 			</div>
 		</div>
 	</div>
